--- a/plugins/cata_donate_utils.pl
+++ b/plugins/cata_donate_utils.pl
@@ -43,8 +43,8 @@ sub AwardEOM {
     if (!$client->GetGM()) {
         quest::set_data($eom_award_log, (quest::get_data($eom_award_log) || 0) + $amount);
         
-        my $act_eom_total_key = $client->AccountID(). " -total-eom-awarded";
-        quest::set_data($act_eom_total_key, (quest::get_data($act_eom_total_key) || 0) + $amount);
+        my $act_eom_total_key = "total-eom-awarded";
+        $client->SetAccountBucket($act_eom_total_key, ($client->GetAccountBucket($act_eom_total_key) || 0) + $amount);
     }
 }
 
@@ -56,8 +56,8 @@ sub AwardEOMAuto {
     if (!$client->GetGM()) {
         quest::set_data($eom_award_log, (quest::get_data($eom_award_log) || 0) + $amount);
 
-        my $act_eom_total_key = $client->AccountID(). " -total-eom-awarded";
-        quest::set_data($act_eom_total_key, (quest::get_data($act_eom_total_key) || 0) + $amount);
+        my $act_eom_total_key = "total-eom-awarded";
+        $client->SetAccountBucket($act_eom_total_key, ($client->GetAccountBucket($act_eom_total_key) || 0) + $amount);
     }
 }