This PR migrates legacy databuckets that generate hundreds of queries per character every time they zone.

This will convert legacy keys such as `Waypoints-account_id` to just `Waypoints` with the `account_id` automatically updated in the data_buckets table.

This will yield only 2 queries when a character loads into a zone versus literally hundreds.

**Backup table before running queries**

**Conversion SQL**

```sql

-- Most things
UPDATE `data_buckets`
SET
  account_id = CAST(SUBSTRING_INDEX(`key`, '-', 1) AS UNSIGNED),
  `key` = SUBSTRING(`key`, LOCATE('-', `key`) + 1)
WHERE
	`key` LIKE '%-ess-items-opened' 
	OR `key` LIKE '%-ess-items-destroyed'
	OR `key` LIKE '%-xmas24_ret_nice'
	OR `key` LIKE '%-xmas24_ret_nice-rewarded'
	OR `key` LIKE '%-progress-flag-%'
	OR `key` LIKE '%CheaterFlag%'
	OR `key` LIKE '%-WarpCount%'
	OR `key` LIKE '%-titles-unlocked'
	OR `key` LIKE '%-group-ports-enabled'
	OR `key` LIKE '%-ess-items-returned'
;

-- Waypoints
UPDATE `data_buckets`
SET
  account_id = CAST(SUBSTRING_INDEX(`key`, '-', -1) AS UNSIGNED),
  `key` = LEFT(`key`, LENGTH(`key`) - LENGTH(SUBSTRING_INDEX(`key`, '-', -1)) - 1)
WHERE `key` LIKE 'Waypoints-%';

-- EOM
UPDATE `data_buckets`
SET
  account_id = CAST(SUBSTRING_INDEX(TRIM(`key`), ' ', 1) AS UNSIGNED),
  `key` = SUBSTRING(SUBSTRING_INDEX(TRIM(`key`), ' ', -1), 2)
WHERE `key` LIKE '% -total-eom-awarded';

UPDATE `data_buckets` SET `key` = 'total-eom-awarded' WHERE `key` = '-total-eom-awarded';
```

**Convert Mis-Cased Max Level**

```sql
update data_buckets set `key` = 'CharMaxLevel' WHERE `key` = 'CharMaxlevel';
```

**Cleanup Flags**

```sql
-- Delete old kill flags
DELETE FROM `data_buckets` WHERE `key` LIKE '%-kill-count%';

-- Delete airplane flags with no expiry
DELETE FROM `data_buckets` WHERE `key` LIKE 'airplane-sirran-%' AND `expires` = 0;

-- These account EOM event scale keys look unused
DELETE FROM `data_buckets` WHERE `key` LIKE '%-eom-event-scale';
````

**Validation Query**

Used to check to see what else is left in the table that may not have been converted.

```sql
select * from data_buckets WHERE
	`key` NOT LIKE '%-ess-items-opened' 
	OR `key` NOT LIKE '%-ess-items-destroyed'
	OR `key` NOT LIKE '%-xmas24_ret_nice'
	OR `key` NOT LIKE '%-xmas24_ret_nice-rewarded'
	OR `key` NOT LIKE '%-progress-flag-%'
	OR `key` NOT LIKE '%CheaterFlag%'
	OR `key` NOT LIKE '%-WarpCount%'
	OR `key` NOT LIKE '%-titles-unlocked'
	OR `key` NOT LIKE '%kill-count%'
	OR `key` NOT LIKE '%Waypoints%'
	OR `key` NOT LIKE '%total-eom-awarded%'
	OR `key` NOT LIKE '%group-ports%'
	OR `key` NOT LIKE '%-ess-items-returned'
	GROUP by `key`;
```

**Example of Logging Into a Zone (Before)**

```
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'CharMaxLevel' LIMIT 1 -- (0 rows returned) (0.000070s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase DELETE FROM data_buckets WHERE character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '0-dev-tools-disabled' -- (0 rows affected) (0.000170s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT `value` FROM `data_buckets` WHERE `key` = 'GestaltClasses' AND `character_id` = 43458 -- (1 row returned) (0.000014s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE account_id IN (43458) AND (`expires` > 1744178493 OR `expires` = 0) -- (0 rows returned) (0.000088s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE character_id IN (43458) AND (`expires` > 1744178493 OR `expires` = 0) -- (103 rows returned) (0.000203s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'CharMaxLevel' LIMIT 1 -- (1 row returned) (0.000014s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT c.`id`, c.`name`, c.`class`, c.`level`, c.`last_login`, c.`zone_id`, g.`guild_id`, g.`rank`, g.`tribute_enable`, g.`total_tribute`, g.`last_tribute`, g.`banker`, g.`public_note`, g.`alt`, g.`online`, db.`value` AS `class_bitmask`  FROM `character_data` AS c  LEFT JOIN `guild_members` AS g ON c.`id` = g.`char_id`  LEFT JOIN `data_buckets` AS db ON c.`id` = db.`character_id` AND db.`key` = 'GestaltClasses'  WHERE c.id=43458 AND c.deleted_at IS NULL -- (1 row returned) (0.000113s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'attack_mode' LIMIT 1 -- (0 rows returned) (0.000035s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-dev-tools-disabled' LIMIT 1 -- (0 rows returned) (0.000015s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE character_id = 43458 AND `key` LIKE 'aaTimer_%%' -- (88 rows returned) (0.000123s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT c.`id`, c.`name`, c.`class`, c.`level`, c.`last_login`, c.`zone_id`, g.`guild_id`, g.`rank`, g.`tribute_enable`, g.`total_tribute`, g.`last_tribute`, g.`banker`, g.`public_note`, g.`alt`, g.`online`, db.`value` AS `class_bitmask`  FROM `character_data` AS c  LEFT JOIN `guild_members` AS g ON c.`id` = g.`char_id`  LEFT JOIN `data_buckets` AS db ON c.`id` = db.`character_id` AND db.`key` = 'GestaltClasses'  WHERE c.id=43458 AND c.deleted_at IS NULL -- (1 row returned) (0.000084s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT c.`id`, c.`name`, c.`class`, c.`level`, c.`last_login`, c.`zone_id`, g.`guild_id`, g.`rank`, g.`tribute_enable`, g.`total_tribute`, g.`last_tribute`, g.`banker`, g.`public_note`, g.`alt`, g.`online`, db.`value` AS `class_bitmask`  FROM `character_data` AS c  LEFT JOIN `guild_members` AS g ON c.`id` = g.`char_id`  LEFT JOIN `data_buckets` AS db ON c.`id` = db.`character_id` AND db.`key` = 'GestaltClasses'  WHERE g.guild_id=199 AND c.deleted_at IS NULL -- (65 rows returned) (0.001949s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'flag-semaphore' LIMIT 1 -- (0 rows returned) (0.000035s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-titles-unlocked' LIMIT 1 -- (1 row returned) (0.000035s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'CharMaxLevel' LIMIT 1 -- (1 row returned) (0.000030s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'CharMaxLevel' LIMIT 1 -- (1 row returned) (0.000041s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-progress-flag-RoK' LIMIT 1 -- (0 rows returned) (0.000033s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-progress-flag-PoP' LIMIT 1 -- (0 rows returned) (0.000030s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'CharMaxLevel' LIMIT 1 -- (1 row returned) (0.000029s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-event-xmas24_ret_nice' LIMIT 1 -- (0 rows returned) (0.000038s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-event-xmas24_ret_nice-rewarded' LIMIT 1 -- (0 rows returned) (0.000014s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-ess-items-destroyed' LIMIT 1 -- (0 rows returned) (0.000025s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-ess-items-opened' LIMIT 1 -- (0 rows returned) (0.000050s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-progress-flag-PoP' LIMIT 1 -- (0 rows returned) (0.000035s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-CheaterFlag' LIMIT 1 -- (0 rows returned) (0.000049s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-titles-unlocked' LIMIT 1 -- (1 row returned) (0.000033s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'CharMaxLevel' LIMIT 1 -- (1 row returned) (0.000029s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'CharMaxLevel' LIMIT 1 -- (1 row returned) (0.000027s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-progress-flag-RoK' LIMIT 1 -- (0 rows returned) (0.000030s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-progress-flag-PoP' LIMIT 1 -- (0 rows returned) (0.000029s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'CharMaxLevel' LIMIT 1 -- (1 row returned) (0.000027s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-event-xmas24_ret_nice' LIMIT 1 -- (0 rows returned) (0.000041s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-event-xmas24_ret_nice-rewarded' LIMIT 1 -- (0 rows returned) (0.000031s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-ess-items-destroyed' LIMIT 1 -- (0 rows returned) (0.000031s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-ess-items-opened' LIMIT 1 -- (0 rows returned) (0.000031s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = '2454-progress-flag-PoP' LIMIT 1 -- (0 rows returned) (0.000019s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 0 AND account_id = 2454 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'event' LIMIT 1 -- (1 row returned) (0.000014s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'PetNameChangesAllowed' LIMIT 1 -- (0 rows returned) (0.000018s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE  character_id = 43458 AND account_id = 0 AND npc_id = 0 AND bot_id = 0 AND zone_id = 0 AND instance_id = 0 AND `key` = 'name_change_allowed' LIMIT 1 -- (0 rows returned) (0.000012s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT c.`id`, c.`name`, c.`class`, c.`level`, c.`last_login`, c.`zone_id`, g.`guild_id`, g.`rank`, g.`tribute_enable`, g.`total_tribute`, g.`last_tribute`, g.`banker`, g.`public_note`, g.`alt`, g.`online`, db.`value` AS `class_bitmask`  FROM `character_data` AS c  LEFT JOIN `guild_members` AS g ON c.`id` = g.`char_id`  LEFT JOIN `data_buckets` AS db ON c.`id` = db.`character_id` AND db.`key` = 'GestaltClasses'  WHERE c.id=43458 AND c.deleted_at IS NULL -- (1 row returned) (0.000192s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT c.`id`, c.`name`, c.`class`, c.`level`, c.`last_login`, c.`zone_id`, g.`guild_id`, g.`rank`, g.`tribute_enable`, g.`total_tribute`, g.`last_tribute`, g.`banker`, g.`public_note`, g.`alt`, g.`online`, db.`value` AS `class_bitmask`  FROM `character_data` AS c  LEFT JOIN `guild_members` AS g ON c.`id` = g.`char_id`  LEFT JOIN `data_buckets` AS db ON c.`id` = db.`character_id` AND db.`key` = 'GestaltClasses'  WHERE g.guild_id=199 AND c.deleted_at IS NULL -- (65 rows returned) (0.001866s)-- [bazaar] (The Bazaar) inst_id [0]
  Zone |   Query    | QueryDatabase SELECT c.`id`, c.`name`, c.`class`, c.`level`, c.`last_login`, c.`zone_id`, g.`guild_id`, g.`rank`, g.`tribute_enable`, g.`total_tribute`, g.`last_tribute`, g.`banker`, g.`public_note`, g.`alt`, g.`online`, db.`value` AS `class_bitmask`  FROM `character_data` AS c  LEFT JOIN `guild_members` AS g ON c.`id` = g.`char_id`  LEFT JOIN `data_buckets` AS db ON c.`id` = db.`character_id` AND db.`key` = 'GestaltClasses'  WHERE c.id=43458 AND c.deleted_at IS NULL -- (1 row returned) (0.000034s)-- [bazaar] (The Bazaar) inst_id [0]
 World |   Query    | QueryDatabase SELECT c.`id`, c.`name`, c.`class`, c.`level`, c.`last_login`, c.`zone_id`, g.`guild_id`, g.`rank`, g.`tribute_enable`, g.`total_tribute`, g.`last_tribute`, g.`banker`, g.`public_note`, g.`alt`, g.`online`, db.`value` AS `class_bitmask`  FROM `character_data` AS c  LEFT JOIN `guild_members` AS g ON c.`id` = g.`char_id`  LEFT JOIN `data_buckets` AS db ON c.`id` = db.`character_id` AND db.`key` = 'GestaltClasses'  WHERE c.id=43458 AND c.deleted_at IS NULL -- (1 row returned) (0.000051s)
  Zone |   Query    | QueryDatabase SELECT id, `key`, value, expires, account_id, character_id, npc_id, bot_id, zone_id, instance_id FROM data_buckets WHERE character_id = 43458 AND `key` LIKE 'aaTimer_%%' -- (88 rows returned) (0.000037s)-- [bazaar] (The Bazaar) inst_id [0]
```

**Other Thoughts**

Consider removing `plugin::CommonCharacterUpdate` from **ENTERZONE** and leaving it for **CONNECT**